requirements.txt
Django==3.2.8
djangorestframework==3.12.4
psycopg2==2.9.1
